[C] Migrate TOC to styled components
