+++ client/src/theme/styles/components/reader/index.js
@@ -12,8 +12,6 @@ import sectionCategoryLabel from "./sectionCategoryLabel";
 import sectionPagination from "./sectionPagination";
 import sectionNextSection from "./sectionNextSection";
 import returnMenu from "./returnMenu";
-import tableOfContents from "./tableOfContents";
-import tocDrawer from "./tocDrawer";
 
 export default `
 ${annotations}
@@ -30,6 +28,4 @@ ${sectionCategoryLabel}
 ${sectionPagination}
 ${sectionNextSection}
 ${returnMenu}
-${tableOfContents}
-${tocDrawer}
 `;
